requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
